% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-confederations.R
\docType{data}
\name{confederations}
\alias{confederations}
\title{FIFA Confederations Data}
\format{
A data frame with 7 rows and 2 variables. Each row represents a FIFA confederation, including FIFA itself.
\describe{
  \item{confederation}{name of the FIFA member association}
  \item{region}{region or continent overseed by the confederation}
}
}
\source{
(2026). \url{https://inside.fifa.com/associations/}.
}
\usage{
confederations
}
\description{
The data was obtained from FIFA website \url{https://inside.fifa.com/associations/} and contains information on the FIFA Member Associations (MAs), also known are confederations, which are responsible for the development and governance of football/soccer within their region
}
\keyword{datasets}
