% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-country-rank.R
\docType{data}
\name{country_rank}
\alias{country_rank}
\title{FIFA World Cup 2026 Country and Ranking Data}
\format{
A data frame with 7 rows and 3 variables. Each row represents a country men's soccer team that qualified for the FIFA World Cup 2026.
\describe{
  \item{country}{name of the country soccer team}
  \item{fifa_rank}{FIFA world ranking as of April, 2026}
  \item{confederation}{region/continent association affiliated to FIFA}
}
}
\source{
Whereig editors (2026). \url{https://www.whereig.com/football/fifa-world-rankings.html/}.
}
\usage{
country_rank
}
\description{
The data was scraped from Whereig website \url{https://www.whereig.com/football/fifa-world-rankings.html/} and contains information on the Men's FIFA World Cup 2026 qualifying teams and ranking data as of 01 April 2026
}
\keyword{datasets}
