% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mx-us-wc-ranks.R
\docType{data}
\name{mx_us_wc_ranks}
\alias{mx_us_wc_ranks}
\title{Mexico's and the United States' World Cup Participation Rankings Data}
\format{
A data frame with 2 rows and 5 variables. Each row represents Mexico's and the United States' men soccer team and their final participation ranking at the last four world cup tournaments.
\describe{
  \item{country}{name of the country soccer team}
  \item{2010}{final ranking in the 2010 world cup}
  \item{2014}{final ranking in the 2014 world cup}
  \item{2018}{final ranking in the 2018 world cup}
  \item{2022}{final ranking in the 2022 world cup}
}
}
\source{
(2026). \url{https://www.thesoccerworldcups.com/world_cups.php}.
}
\usage{
mx_us_wc_ranks
}
\description{
The data was gathered from The Soccer World Cups website \url{https://www.thesoccerworldcups.com/world_cups.php} and contains information about every World Cup played including national teams, standings, and more
}
\keyword{datasets}
