% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-country-capital.R
\docType{data}
\name{country_capital}
\alias{country_capital}
\title{FIFA World Cup 2026 Country and Capital Data}
\format{
A data frame with 5 rows and 3 variables. Each row represents a country men's soccer team that qualified for the FIFA World Cup 2026.
\describe{
  \item{country}{name of the country}
  \item{capital}{name of the country's capital city}
  \item{population}{population size in millions based on the United Nations Population Division estimates for 2025}
}
}
\source{
(2026). \url{https://theworlddata.com/world-population-by-country/}.

(2025). \url{https://www.worlddata.info/capital-cities.php/}.
}
\usage{
country_capital
}
\description{
The data was obtained from The World Data website \url{https://theworlddata.com/world-population-by-country/} which contains information on the Men's FIFA World Cup 2026 qualifying teams and their 2025 population, 
and from the WorldData.info website \url{https://www.worlddata.info/capital-cities.php/} which contains information on the capital cities of all countries.
}
\keyword{datasets}
